+++ src/utils/wait.cpp
@@ -1,6 +1,6 @@
 #include "utils/utils.h"
 
-void Timer::setTimer(unsigned int ms)
+void Timer::setTimer(unsigned long ms)
 {
     this->end = millis() + ms;
 };
